+++ src/graphql/schema/create_entities/create_resolver/create_resolver_input_value/mod.rs
@@ -5,7 +5,7 @@ use crate::{
         schema::{ExcludeFromInput, ResolverType, ServiceSchemaBuilder},
     },
 };
-use async_graphql::dynamic::{Field, InputValue, TypeRef};
+use async_graphql::dynamic::{Field, InputObject, InputValue, TypeRef};
 use log::debug;
 
 mod get_resolver_input_name;
@@ -20,42 +20,22 @@ impl ServiceSchemaBuilder {
     ) -> Self {
         debug!("Creating Resolver Input");
 
+        let mut inputs = Vec::new();
+
         let input_name =
             ServiceSchemaBuilder::get_resolver_input_name(&entity.name, resolver_type, None);
 
-        let exclude_from_input = match resolver_type {
-            ResolverType::FindOne => Some(ExcludeFromInput::FindOne),
-            ResolverType::CreateOne => Some(ExcludeFromInput::CreateOne),
-            ResolverType::FindMany => Some(ExcludeFromInput::FindMany),
-            ResolverType::UpdateOne => Some(ExcludeFromInput::UpdateOne),
-            ResolverType::UpdateMany => Some(ExcludeFromInput::UpdateMany),
-            ResolverType::InternalType => None,
-        };
+        let mut root_input = InputObject::new(&input_name);
 
-        let mut inputs = ServiceInput::new(
-            input_name.clone(),
-            entity.fields.clone(),
-            resolver_type.clone(),
-            exclude_from_input,
-        )
-        .build(Some(true));
-
-        let include_query_input =
-            resolver_type == &ResolverType::UpdateOne || resolver_type == &ResolverType::UpdateMany;
+        let include_query_input = resolver_type == &ResolverType::UpdateOne
+            || resolver_type == &ResolverType::UpdateMany
+            || resolver_type == &ResolverType::FindOne
+            || resolver_type == &ResolverType::FindMany;
 
         if include_query_input {
-            let mut query_input = match inputs
-                .iter()
-                .position(|input| input.type_name() == input_name)
-                .map(|i| inputs.remove(i))
-            {
-                Some(input) => input,
-                None => return self,
-            };
-
             let query_input_name = match resolver_type {
                 ResolverType::UpdateOne => ServiceSchemaBuilder::get_resolver_input_name(
-                    &entity.name.to_lowercase(),
+                    &format!("{}_query", &entity.name.to_lowercase()),
                     &ResolverType::FindOne,
                     None,
                 ),
@@ -64,12 +44,24 @@ impl ServiceSchemaBuilder {
                     &ResolverType::UpdateOne,
                     None,
                 ),
+                ResolverType::FindOne => ServiceSchemaBuilder::get_resolver_input_name(
+                    &format!("{}_query", &entity.name.to_lowercase()),
+                    &ResolverType::FindOne,
+                    None,
+                ),
+                ResolverType::FindMany => ServiceSchemaBuilder::get_resolver_input_name(
+                    &format!("{}s_query", &entity.name.to_lowercase()),
+                    &ResolverType::FindMany,
+                    None,
+                ),
                 _ => unreachable!(),
             };
 
             let exclude_from_input = match resolver_type {
                 ResolverType::UpdateOne => Some(ExcludeFromInput::UpdateOneQuery),
                 ResolverType::UpdateMany => Some(ExcludeFromInput::UpdateManyQuery),
+                ResolverType::FindOne => Some(ExcludeFromInput::FindOne),
+                ResolverType::FindMany => Some(ExcludeFromInput::FindMany),
                 _ => unreachable!(),
             };
 
@@ -81,14 +73,62 @@ impl ServiceSchemaBuilder {
             )
             .build(Some(true));
 
-            query_input = query_input.field(InputValue::new(
+            root_input = root_input.field(InputValue::new(
                 "query",
                 TypeRef::named_nn(query_input_name.clone()),
             ));
 
-            inputs.push(query_input);
             inputs.extend(rest_inputs);
         }
+
+        let include_values_input = resolver_type == &ResolverType::CreateOne
+            || resolver_type == &ResolverType::UpdateOne
+            || resolver_type == &ResolverType::UpdateMany;
+
+        if include_values_input {
+            let values_input_name = match resolver_type {
+                ResolverType::CreateOne => ServiceSchemaBuilder::get_resolver_input_name(
+                    &format!("{}_values", &entity.name.to_lowercase()),
+                    &ResolverType::CreateOne,
+                    None,
+                ),
+                ResolverType::UpdateOne => ServiceSchemaBuilder::get_resolver_input_name(
+                    &format!("{}_values", &entity.name.to_lowercase()),
+                    &ResolverType::UpdateOne,
+                    None,
+                ),
+                ResolverType::UpdateMany => ServiceSchemaBuilder::get_resolver_input_name(
+                    &format!("{}s_values", &entity.name.to_lowercase()),
+                    &ResolverType::UpdateOne,
+                    None,
+                ),
+                _ => unreachable!(),
+            };
+
+            let exclude_from_input = match resolver_type {
+                ResolverType::CreateOne => Some(ExcludeFromInput::CreateOne),
+                ResolverType::UpdateOne => Some(ExcludeFromInput::UpdateOne),
+                ResolverType::UpdateMany => Some(ExcludeFromInput::UpdateMany),
+                _ => unreachable!(),
+            };
+
+            let rest_inputs = ServiceInput::new(
+                values_input_name.clone(),
+                entity.fields.clone(),
+                ResolverType::FindOne,
+                exclude_from_input,
+            )
+            .build(None);
+
+            root_input = root_input.field(InputValue::new(
+                "values",
+                TypeRef::named_nn(values_input_name.clone()),
+            ));
+
+            inputs.extend(rest_inputs);
+        }
+
+        inputs.push(root_input);
 
         if !inputs.is_empty() {
             resolver = resolver.argument(InputValue::new(